Swapping affects the BattlePowerLoadout objects appropriately. Aesthetic changes made. New bug found
Added and used new commands to swap the cards' loadouts after the on-screen swapping is complete.

Made it so that after the cards are initially loaded, the layout groups are disabled so there's no awkward card-rearranging when you start dragging a card.

New bug: When you try to swap cards between the left side of the second row in storage,  and the right side of the first row, the card you last hovered over stays in place instead of switching with the card you were dragging.