Fixed bug with cards inappropriately staying in place. New one found
The trigger conditions are the same as with the previous bug. This time, instead of any cards getting stuck in the wrong place, the card you finished dragging just doesn't go back to normal size until you drag and swap it again.

While it's still too big, other objects hovering over it can't swap with it.